requirements.txt
asyncio-nats-client==0.10.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
future-fstrings==1.2.0
idna==3.3
importlib-metadata==4.11.3
influxdb==5.2.3
jsonschema==3.1.1
lxml==4.8.0
nats-python==0.8.0
numpy==1.22.3
ordered-set==4.1.0
pandas==1.4.1
pyecore==0.12.2
pyrsistent==0.18.1
python-dateutil==2.8.0
pytz==2022.1
pywin32==303
requests==2.27.1
RestrictedPython==5.2
six==1.16.0
urllib3==1.26.9
zipp==3.7.0
pyESDL=21.12.1
